pyproject.toml
[project]
name = "estagio-teste"
version = "0.1.0"
description = ""
authors = [
    {name = "T-Terra",email = "gabst138@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "selenium (>=4.30.0,<5.0.0)",
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "dj-database-url (>=2.3.0,<3.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "tabula-py (>=2.10.0,<3.0.0)",
    "jpype1 (>=1.5.2,<2.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.3.5"
pytest-django = "^4.10.0"

[tool.pytest.ini_options]
python_files = ["test_*.py", "*_test.py", "testing/python/*.py", "tests/*.py"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
